Cleanup Round 2
* All sub-commands get their own executable -- keeps vip.js easier to read and maintain
* Use const for all module requires
* Include as much usage information as possible in command definitions

Also: Deprecate --export for `vip db`. You can still use export and
shouldn't notice a difference unless you're trying to see the full
database in your terminal (probably not that useful). Instead, we detect
when stdout is piped and assume you want a mysqldump.
